+++ Lab2/iterator.py
@@ -1,66 +1,71 @@
-import argparse
 import csv
 
 
-class DatasetIterator:
-    def __init__(self, annotation_file: str, class_name: str):
+class ImageIterator:
+    def __init__(self, annotation_file, target_class):
         self.annotation_file = annotation_file
-        self.class_name = class_name
-        self.class_instances = None
-        self.current_index = 0
+        self.target_class = target_class
+        self.class_instances = self.load_annotation_file()
+        self.index = 0
+        self.prev_instances = set()
 
-    def __iter__(self):
-        self.class_instances = self._get_class_instances()
-        return self
-
-    def _get_class_instances(self):
+    def load_annotation_file(self):
         instances = []
         with open(self.annotation_file, 'r') as csvfile:
             csvreader = csv.reader(csvfile)
             for row in csvreader:
-                if len(row) == 3 and row[2] == self.class_name:
+                if len(row) == 3 and row[2] == self.target_class:
                     instances.append(row[0])
         return instances
-
-    def __next__(self):
-        if self.current_index < len(self.class_instances):
-            next_instance = self.class_instances[self.current_index]
-            self.current_index += 1
-            return next_instance
-        else:
-            raise StopIteration
-
-
-class ClassIterator:
-    def __init__(self, annotation_file: str, class_names: list):
-        self.annotation_file = annotation_file
-        self.class_names = class_names
-        self.current_name_index = 0
 
     def __iter__(self):
         return self
 
     def __next__(self):
-        if self.current_name_index < len(self.class_names):
-            current_class_name = self.class_names[self.current_name_index]
-            self.current_name_index += 1
-            return DatasetIterator(self.annotation_file, current_class_name)
-        else:
-            raise StopIteration
+        while self.index < len(self.class_instances):
+            instance = self.class_instances[self.index]
+            self.index += 1
+            if instance not in self.prev_instances:
+                self.prev_instances.add(instance)
+                return instance
+
+        raise StopIteration("No more instances for the specified class.")
+
+
+class ClassIterator:
+    def __init__(self, csv_file: str, class_labels):
+        self.csv_file = csv_file
+        self.class_labels = class_labels
+        self.image_iterators = {label: ImageIterator(csv_file, label) for label in class_labels}
+        self.current_label_index = 0
+        self.current_image_iterator = None
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        if not self.class_labels or self.current_label_index >= len(self.class_labels):
+            raise StopIteration("No more instances for the specified class.")
+
+        current_class_label = self.class_labels[self.current_label_index]
+
+        if self.current_image_iterator is None:
+            self.current_image_iterator = self.image_iterators[current_class_label]
+
+        try:
+            return next(self.current_image_iterator)
+        except StopIteration:
+            self.current_label_index += 1
+            self.current_image_iterator = None
+            return next(self)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description='Process a dataset')
-    parser.add_argument('annotation_file', type=str, help='Path to annotation')
-    parser.add_argument('class_names', type=str, nargs='+', help='List of class names')
+    annotation_file = 'annotation.csv'
+    class_labels = ["cats", "dogs"]
+    iterator = ClassIterator(annotation_file, class_labels)
 
-    args = parser.parse_args()
-
-    annotation_file = args.annotation_file
-    class_names = args.class_names
-
-    iterator = ClassIterator(annotation_file, class_names)
-
-    for dataset_iterator in iterator:
-        for instance in dataset_iterator:
-            print(instance)
+    instance = iterator.next_image()
+    while instance is not None:
+        print(instance)
+        instance = iterator.next_image()
